add-uv (#62)
* Fixing dependencies

* fixed typo

* Update setup.py with the correct dependency for imbalanced-learn

Without this change an automatic dependency licenses check would fail because imblearn does not a licence, the name of the package is imbalanced-learn.

* Updated version

* Changed tests folder name
added pyproject.toml
added uv
added .venv to .gitignore

* Fixing dependencies

* fixed typo

* Update setup.py with the correct dependency for imbalanced-learn

Without this change an automatic dependency licenses check would fail because imblearn does not a licence, the name of the package is imbalanced-learn.

* Updated version

* Changed tests folder name
added pyproject.toml
added uv
added .venv to .gitignore

* changing building pipelines

* changing building pipelines

* pumping versions of pandas

* fixing and disabling some tests

* fixing CI

* fixing CI

* adding cov xdist

* desabling failing tests

* fixing pipeline

* fixing pipeline

* disabling circleCI

---------

Co-authored-by: Anastasia Korobkina <a.korobkina.19@gmail.com>
Co-authored-by: Emanuele Albini <7670445+emanuele-albini@users.noreply.github.com>